Move shipment methods out of Order and into OrderComponents::Shipment
